+++ src/main/java/com/hwmods/overpressure/Overpressure.java
@@ -31,6 +31,9 @@ public class Overpressure {
         modBus.addListener(this::commonSetup);
         modBus.addListener(TubePlacementSyncPayload::register);
         modBus.addListener(TubePlacementReachPayload::register);
+        modBus.addListener(com.hwmods.overpressure.tube.TubeSectionPayload::register);
+        modBus.addListener(com.hwmods.overpressure.tube.TubeSectionInteractionPayload::register);
+        modBus.addListener(com.hwmods.overpressure.tube.SectionCargoPayload::register);
         modContainer.registerConfig(ModConfig.Type.COMMON, Config.SPEC);
     }